v1.2.0
- added carescript, a DSL to customize the downloading process
- added option "blacklist"
- added option "template"
- added config option "show_script_src"
- added config option "no_scripts"
- option "release" not automatically updates the cat_checklist.inipp
- fixed bug with "add" not working correctly

+++ mods/inipp.hpp
@@ -198,7 +198,7 @@ public:
         type = IniType::Int;
         return *this;
     }
-    inline IniElement operator=(float floatp) {
+    inline IniElement operator=(long double floatp) {
         src = std::to_string(floatp);
         type = IniType::Float;
         return *this;
@@ -449,7 +449,7 @@ public:
 
     inline void set(std::string key, std::string value, std::string section = "Main") { get(key,section) = value; }
     inline void set(std::string key, int value, std::string section = "Main") { get(key,section) = value; }
-    inline void set(std::string key, float value, std::string section = "Main") { get(key,section) = value; }
+    inline void set(std::string key, long double value, std::string section = "Main") { get(key,section) = value; }
 
     inline void construct(std::string key, std::string source, std::string section = "Main") {
         IniFile::set(key,IniElement(source),section);
